refactor(inference): extract cloud tests to sibling files
primer-inference/src/cloud.rs carried two sibling trailing test modules
(`mod tests` and `mod classify_tests`), pushing it to 938 lines. Move each
into its own sibling file under src/cloud/, declared as `mod tests;` /
`mod classify_tests;` from cloud.rs.

Splitting into two separate sibling files (rather than nesting both under one
cloud/tests.rs) keeps each `use super::*` resolving to `cloud` verbatim — no
glob-of-glob trap, zero visibility churn, no signature change. cloud.rs drops
938 -> 480 lines (under the ~500 guideline). All 37 cloud tests (21 + 16)
preserved at their exact module paths and passing.

Verified: cargo test -p primer-inference --lib (199 ok), cargo clippy
-p primer-inference --all-targets -D warnings clean, cargo fmt --all --check
clean, cargo test --workspace (0 failures).
